Script to recover intact fastq reads
Script to recover reads from a fastq file which may be corrupted and might have stuff that are not fastq reads in it. It checks the integrity of the reads:  The id line contain a valid id? The sequence is only nucleotides?  The comment line is just a small comment and starts with '+'? The quality has the same size as the nucleotide sequence? If it pass, the read is printed, else, it is skipped.